+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
 ## Unreleased
 
+### Added
+
+- Import and export of Systems and Branches
+
 ### Fixed
 
 - Avatar imports now work if the avatars aren't in the root of the ZIP file